internal/tasks/tasks.go
// tasks.go contains the logic for the distributed tasks that the MPI-like framework
// executes across multiple nodes. Each task is broken down into smaller parts
// and distributed among the worker nodes, with results communicated via MPI functions.
package tasks

import (
	"fmt"
)

// DistributeTask distributes the task (e.g., data) to the nodes
func DistributeTask(data []int) error {
	// Logic to distribute tasks across nodes
	fmt.Println("Distributing task to nodes")
	return nil
}

// ProcessTask processes the task on a specific node (e.g., sorting or computation)
func ProcessTask(taskID int, data []int) ([]int, error) {
	// Logic to process the task (e.g., sorting the array)
	fmt.Printf("Processing task %d on node\n", taskID)
	return data, nil
}

// CollectResults collects the results from the nodes after the task has been processed
func CollectResults() ([]int, error) {
	// Logic to collect results from all nodes
	fmt.Println("Collecting results from nodes")
	return []int{1, 2, 3}, nil
}
